+++ js/bjcp-to-json/dataFlattenFixed.json
@@ -14710,7 +14710,7 @@
   "2240.items.5.text": "entering",
   "2240.items.6.text": " a beer in one of these styles.",
   "2241.lineBreak": true,
-  "2242.items.0.text": "34A. Commercial Specialty Beer",
+  "2242.text": "34A. Commercial Specialty Beer",
   "2243.items.0.text": "This style is intended for reproductions",
   "2243.items.1.text": " or interpretations",
   "2243.items.2.text": " of specific commercial beers that ",
@@ -15191,7 +15191,7 @@
   "2964.items.0.text": "Italian",
   "2964.items.1.format.color": "#1F4E79",
   "2964.items.1.text": " Styles",
-  "2965.items.0.text": "X3. Italian Grape Ale",
+  "2965.text": "X3. Italian Grape Ale",
   "2966.items.0.text": "For uses outside Italy, see",
   "2966.items.1.text": " 29D Grape Ale.",
   "2967.items.0.format.bold": true,
@@ -15292,7 +15292,7 @@
   "2980.items.0.text": "Brazilian",
   "2980.items.1.format.color": "#1F4E79",
   "2980.items.1.text": " Styles",
-  "2981.items.0.text": "X4. Catharina Sour",
+  "2981.text": "X4. Catharina Sour",
   "2982.format.bold": true,
   "2982.format.italic": true,
   "2982.format.size": "18",
@@ -15409,7 +15409,7 @@
   "2993.format.textAlign": "left",
   "2993.items.0.format.bold": true,
   "2993.items.0.format.size": "18",
-  "2993.items.0.text": "IBU:",
+  "2993.items.0.text": "IBUs:",
   "2993.items.1.format.bold": true,
   "2993.items.1.format.size": "18",
   "2993.items.1.text": "2 – 8",
@@ -15649,7 +15649,7 @@
   "3016.items.0.text": "New Zealand",
   "3016.items.1.format.color": "#1F4E79",
   "3016.items.1.text": " Styles",
-  "3017.items.0.text": "X5. New Zealand Pilsner",
+  "3017.text": "X5. New Zealand Pilsner",
   "3018.items.0.text": "Suggested style placement: Category ",
   "3018.items.1.text": "12",
   "3018.items.2.text": " (",
