+++ js/bjcp-to-json/styles/X4. Catharina Sour.json
@@ -13,7 +13,6 @@
   "historia": "Os exemplos individuais existiam com nomes diferentes anteriormente no Brasil, mas o estilo se tornou popular com esse nome depois que foi definido formalmente em 2015 durante uma reunião entre cervejeiros profissionais e caseiros em Santa Catarina. Utilizando ingredientes locais adequados para um clima quente, o estilo se espalhou para outros estados do Brasil e além, sendo um estilo muito popular na América do Sul – tanto em competições comerciais como caseiras.",
   "ingredientes": "Malte Pilsen com malte de trigo ou trigo não maltado. A técnica de Kettle Sour com o uso de Lacto é a mais comum de ser utilizada, seguida por uma fermentação com uma levedura ale neutra. A fruta é tipicamente adicionada nos estágios finais da fermentação. Frutas da estação frescas, comumente tropicais. Especiarias, ervas e vegetais são opcionais, mas devem sempre estar em caráter de apoio e elevar a percepção da fruta.",
   "comparacoes_de_estilo": "Como uma Berliner Weisse mais forte, mas com fruta fresca e sem Brett. Menos ácida do que Lambic e Gueuze e sem o caráter da Brett. A partir do guia de estilos 2021, cervejas semelhantes podem ser inscritas no estilo mais amplo 28C Wild Specialty Beer Style.",
-  "ibu": "2 – 8",
   "exemplos_comerciais": "Armada Daenerys, Blumenau Catharina Sour Pêssego, Istepô Goiabêra, Itajahy Catharina Araçá Sour, Liffey Coroa Real, UNIKA Tangerina Clemenules",
   "marcacoes": "estilo-craft, fruta, ácida, cerveja-specialty",
   "original_gravity": {
@@ -24,6 +23,16 @@
     "maximum": {
       "unit": "sg",
       "value": 1.048
+    }
+  },
+  "international_bitterness_units": {
+    "minimum": {
+      "unit": "IBUs",
+      "value": 2
+    },
+    "maximum": {
+      "unit": "IBUs",
+      "value": 8
     }
   },
   "final_gravity": {
